Add and ship file that lists the SDK versions XI supports.
Add and ship a file that lists the SDK versions of the various SDKs XI supports.

Also list the minimum SDK version for each extension.

And add a script that verifies that the current SDK version is in this file,
which should ensure every SDK version bump ends up in the file.