Add and ship file that lists the SDK versions XI/XM supports.
* Add and ship file that lists the SDK versions XI supports.

Add and ship a file that lists the SDK versions of the various SDKs XI supports.

Also list the minimum SDK version for each extension.

And add a script that verifies that the current SDK version is in this file,
which should ensure every SDK version bump ends up in the file.

* [mtouch] Remove generated file from source control.

* [SdkVersions] Remove versions not in master yet.

* [SdkVersions] Improve detection script to verify that versions are between min and max, and that both min and max are in the list.

* Move things around a bit and add macOS support.

* Update Versions-mac.plist.in with extension info