[CoreVideo] Fix memory leak in CVPixelBuffer.Create for planar bytes.
When creating a CVPixelBuffer with planar bytes, we create one GCHandle for
every byte[] of planar data, as well as one GCHandle for a custom object that
has an array of all the other GCHandles.

All these GCHandles were freed properly if the CFPixelBuffer was successfully
created, but in the case of failure, only the GCHandle for the custom object
was freed.

So make sure to free all the GCHandles even in the case of failures by calling
the free callback function.