[foundation] Fix possible abort when using NSUrlSessionHandler. Fix #50891 (#1451) (#1452)
PR 1435 did not fix bug #50891. The default (native) implementation till
get called and abort the process. The right way to solve this is to
override the members (easier now that it's not internal anymore) and
return `false` (or `null`) so the caller (the OS) knows it's not
supported (that's fine)

references:
* https://bugzilla.xamarin.com/show_bug.cgi?id=50891
* https://github.com/xamarin/xamarin-macios/pull/1435