[xharness] Improve handling and reporting of unexpected exceptions.
Unexpected exceptions are now caught in a more central location
(TestTask.RunInternalAsync), which means it's not necessary to catch
unexpected exceptions in other places (so those catch handlers have been
removed).

Additionally we now set the failure message when such an exception occurs, in
addition to writing the full exception details to a custom log file.