[bmac] Parse --new-style again.
This was changed in 1b1e345d7259 since bmac.exe says --new-style is
deprecated.

However some users (MonoTouch.Hosting) manually call bmac (instead of going
through a binding project), and with 1b1e345d7259 we changed the behavior when
--new-style was passed, so that it ended up trying to build Classic instead
(which is clearly not right for --new-style).

So make the bmac script understand --new-style again (to not break
compatibility), but don't forward the flag to bmac.exe (to avoid the
deprecation warning).