[tests][intro] Check if some API should have an *Async version provided
Taken out of our Xcode8 post-mortem: it's easy to forget (while adding)
or miss (while auditing) a `[Async]` attribute on the API that would
benefit from them.

Not meant to be merged (just yet). The list of failures (per platform)
needs to be reviewed to see if this generates too many false positives
and becomes a time burner instead of a time saver.