[generator] Allow the use of [Async] attribute on [Abstract] decorated methods. Fixes #43165
Fixing the compilation was not enough [1] as it would still require to implement the async variant when subclassing. That could already be done outside the bindings (no saving).

The good news is that the generated `*Async` method does not need to be `abstract` even if it calls an `abstract` method. We already have enough information (from the `abstract` signature) to generated the async wrapper method.

Ideally (too late for the current profile) the `*Async` method should not be virtual, the async wrapper does not need to be overridden, just the binding call -> XAMCORE_4_0.

reference:
https://bugzilla.xamarin.com/show_bug.cgi?id=43165

[1] it was missing a `;` - but also caused other compilation issues