[generator] Fix a potential compile time error in generated *Async methods
This can happen when a parameter of the method being [Async]-ified has
the same name as one of the delegate parameter.

E.g.

* Binding code

```
	delegate void MTKTextureLoaderCallback ([NullAllowed] IMTLTexture texture, [NullAllowed] NSError error);

	[Async]
	void FromTexture (MDLTexture texture, [NullAllowed] MTKTextureLoaderOptions options, MTKTextureLoaderCallback completionHandler);
```

* Generated code

```
		public unsafe virtual Task<global::Metal.IMTLTexture> FromTextureAsync (global::ModelIO.MDLTexture texture, NSDictionary options)
		{
			var tcs = new TaskCompletionSource<global::Metal.IMTLTexture> ();
>			FromTexture(texture, options, (texture, error) => {
				if (error != null)
					tcs.SetException (new NSErrorException(error));
				else
					tcs.SetResult (texture);
			});
			return tcs.Task;
		}
```

* Errpr

```
MTKTextureLoader.g.cs(397,35): error CS0136: A local variable named `texture' cannot be declared in this scope because it would give a different meaning to `texture', which is already used in a `parent or current' scope to denote something else
```