Re-enable XM
When building on wrench we need XM (mmp) to build `mlaunch` and
this means have XM enabled - else no package will be produced.

A better (long term) fix would be to have `INCLUDE_MAC` only
control the packaging and tests of XM so the tooling is present
to complete XI builds.