Fix Bug 51514 - [generator] Missing [Wrap] from protocol/interface when parent has a [Wrap] itself (#1708)
https://bugzilla.xamarin.com/show_bug.cgi?id=51514

The generator used a list of strings (selector names) to determine
what methods are needed to be generated. Unfortunately this info
is not available (null) when looking for it on [Wrap] decorated members
so it would only generate the first [Bind] it would find. Now it uses
a HashSet of MemberInformation objects to compare.

The removal of SCNView.HitTest inside scenekit.cs was done because
this is now correctly generated from SCNSceneRenderer interface