[generator] Don't use Type.GetType to look in all loaded assemblies, look directly in the right assembly.
'Type.GetType' doesn't work the same with IKVM as with Reflection, so refactor
to code that works the same (and just as correctly) for both IKVM and
Reflection.

In particular there's no need to use 'Type.GetType' with an assembly qualified
typename when we already have the assembly we need to look in: just use the
type's FullName to look up the type in the assembly instead.

Generator diff: https://gist.github.com/rolfbjarne/a3da6c9df53383c4225483f89e16d156