[generator] Bump .NET version for the generator's project file.
Array.Empty<T> did not exist in earlier versions of .NET, so this would build
fine on the command-line (which builds using the system's latest .NET
version), but not with the project file (which apparently defaults to .NET 4.5
unless otherwise specified).

So bump the .NET version in the project file so that compilation works the
same in both cases.