[Proposal] Add playground to IKVM generator template
When adding new features to the generator we find ourselves
creating new test cases in test/generator or with binding projects
and not actually modifying the actual SDK api definition source files
because you get a lot of noise on the generated code.

By doing the above it would not allow us to leverage the new debugging
capabilities (thanks Rolf!!) that the new IKVM based generator grant us.

So my proposal is to have this Playground (generator-playground.cs)
where you can add arbitrary interface definitions to test and debug the
generator that sould end up as your test case in tests/generator.