[generator] Fix Bug 53103 - Using [Async] on a method that does no return `void` should be an warning
https://bugzilla.xamarin.com/show_bug.cgi?id=53103

We currently allow using [Async] on members whose return type is not
`System.Void`, this leads to silently ignoring the actual returned
result of the called member.

We have some cases on our own but fixing them is a breaking change
so if XAMCORE_4_0 ever happens we will have another look to them.

We now issue a warning (BI1118) whenever we find `[Async]` being used
on a member whose return type **is not** `System.Void` and if you
really want to have `[Async]` on members with different return type
to `System.Void` you can silence the warning by using
`[Async (allowNonVoidReturnType: true)]`, also using the provided
overloads or by setting the property `AllowNonVoidReturnType` to
`true` of the `[AsyncAttribute]`.

If you use `[Async]` on a member whose return type **is not** `System.Void`
the returned value of the member will be **ignored**.